agent/agent.py
import logging
from language_model import openai_api

logger = logging.getLogger(__name__)

class AutomatedSoftwareEngineer:
    def __init__(self, name, tools, api_key):
        self.name = name
        self.tools = tools
        self.openai_api = openai_api.OpenAIAPI(api_key)
        self.inner_thoughts = []

    def complete_task(self, task_description):
        logger.info(f"Completing task: {task_description}")
        
        # Use OpenAI API to generate code
        generated_code = self.openai_api.complete_task(task_description)
        
        # Add any reflection or additional processing here
        
        return generated_code